requirements.txt
pyspark>=3.5.0
pyyaml>=6.0
great-expectations>=1.0.0
dbt-core>=1.10.0
dbt-snowflake>=1.10.0
snowflake-connector-python>=3.0.0
